Tracker: Added elementary routing of data to Designer
Former-commit-id: b86bf3af3a6173c04c4557aa752c81bf3badf5cd

+++ applications/platform/tracker/include/Dataset.h
@@ -18,9 +18,9 @@ public:
 
 	std::vector<Marker> m_vMarkers;           // Replace with IStimulationSet? 
 	std::vector<OpenViBE::CMatrix*> m_vEEG;    // Signal
-	uint64_t m_samplingRate;
+	uint32_t m_samplingRate;
 	uint64_t m_Duration;
-	uint64_t m_chunkSize;
+	uint32_t m_chunkSize;
 
 	std::string m_Filename;
